+++ kroxylicious-api/src/main/java/io/kroxylicious/proxy/config/secret/FilePassword.java
@@ -13,7 +13,6 @@ import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Objects;
 
-import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
@@ -22,12 +21,10 @@ import edu.umd.cs.findbugs.annotations.NonNull;
  * A reference to the file containing a plain text password in UTF-8 encoding.  If the password file
  * contains more than one line, only the characters of the first line are taken to be the password,
  * excluding the line ending.  Subsequent lines are ignored.
- * <br>
- * Note that the {@code filePath} alias was deprecated at 0.5.0.
  *
  * @param passwordFile file containing the password.
  */
-public record FilePassword(@JsonProperty(required = true) @JsonAlias("filePath") String passwordFile) implements PasswordProvider {
+public record FilePassword(@JsonProperty(required = true) String passwordFile) implements PasswordProvider {
 
     public FilePassword {
         Objects.requireNonNull(passwordFile);