+++ kroxylicious-operator/src/test/java/io/kroxylicious/kubernetes/operator/OperatorTestUtils.java
@@ -13,8 +13,23 @@ import io.fabric8.kubernetes.client.KubernetesClientException;
 import edu.umd.cs.findbugs.annotations.Nullable;
 
 public class OperatorTestUtils {
+
+    /**
+     * The timeouts etc of this client build are tuned to handle the case where Kubernetes isn't present.
+     * As might be the case on a developer's machine where minikube isn't running.
+     */
+    private static final KubernetesClientBuilder PRESENCE_PROBING_KUBE_CLIENT_BUILD = new KubernetesClientBuilder()
+            .editOrNewConfig()
+            .withRequestRetryBackoffLimit(2)
+            .withConnectionTimeout(500)
+            .endConfig();
+
     static @Nullable KubernetesClient kubeClientIfAvailable() {
-        var client = new KubernetesClientBuilder().build();
+        return kubeClientIfAvailable(new KubernetesClientBuilder());
+    }
+
+    static @Nullable KubernetesClient kubeClientIfAvailable(KubernetesClientBuilder kubernetesClientBuilder) {
+        var client = kubernetesClientBuilder.build();
         try {
             client.namespaces().list();
             return client;
@@ -26,7 +41,7 @@ public class OperatorTestUtils {
     }
 
     static boolean isKubeClientAvailable() {
-        try (var client = kubeClientIfAvailable()) {
+        try (var client = kubeClientIfAvailable(PRESENCE_PROBING_KUBE_CLIENT_BUILD)) {
             return client != null;
         }
     }