+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/admin/ManagementInitializer.java
@@ -5,6 +5,9 @@
  */
 package io.kroxylicious.proxy.internal.admin;
 
+import java.util.Objects;
+import java.util.Optional;
+
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelPipeline;
 import io.netty.channel.socket.SocketChannel;
@@ -12,18 +15,24 @@ import io.netty.handler.codec.http.HttpResponseStatus;
 import io.netty.handler.codec.http.HttpServerCodec;
 import io.netty.handler.codec.http.HttpServerExpectContinueHandler;
 
-import io.kroxylicious.proxy.config.admin.AdminHttpConfiguration;
+import io.kroxylicious.proxy.config.admin.EndpointsConfiguration;
+import io.kroxylicious.proxy.config.admin.ManagementConfiguration;
 import io.kroxylicious.proxy.internal.MeterRegistries;
 
-public class AdminHttpInitializer extends ChannelInitializer<SocketChannel> {
+import edu.umd.cs.findbugs.annotations.NonNull;
+
+public class ManagementInitializer extends ChannelInitializer<SocketChannel> {
 
     private static final String LIVEZ = "/livez";
     private final MeterRegistries registries;
-    private final AdminHttpConfiguration adminHttpConfiguration;
+    private final ManagementConfiguration managementConfiguration;
 
-    public AdminHttpInitializer(MeterRegistries registries, AdminHttpConfiguration adminHttpConfiguration) {
+    public ManagementInitializer(@NonNull MeterRegistries registries,
+                                 @NonNull ManagementConfiguration managementConfiguration) {
+        Objects.requireNonNull(registries);
+        Objects.requireNonNull(managementConfiguration);
         this.registries = registries;
-        this.adminHttpConfiguration = adminHttpConfiguration;
+        this.managementConfiguration = managementConfiguration;
     }
 
     @Override
@@ -33,9 +42,9 @@ public class AdminHttpInitializer extends ChannelInitializer<SocketChannel> {
         p.addLast(new HttpServerExpectContinueHandler());
         RoutingHttpServer.RoutingHttpServerBuilder builder = RoutingHttpServer.builder();
         builder.withRoute(LIVEZ, httpRequest -> RoutingHttpServer.responseWithStatus(httpRequest, HttpResponseStatus.OK));
-        adminHttpConfiguration.endpoints().maybePrometheus().ifPresent(prometheusMetricsConfig -> {
-            builder.withRoute(PrometheusMetricsEndpoint.PATH, new PrometheusMetricsEndpoint(registries));
-        });
+        Optional.ofNullable(managementConfiguration.endpoints())
+                .map(EndpointsConfiguration::maybePrometheus)
+                .ifPresent(prometheusMetricsConfig -> builder.withRoute(PrometheusMetricsEndpoint.PATH, new PrometheusMetricsEndpoint(registries)));
         p.addLast(builder.build());
     }